tkdraw/tk: Move tk_elems.py into new tk sub-module.
